demo todo added and set confirm popup  in delete handler

+++ src/Context/TodoContext.jsx
@@ -4,37 +4,52 @@ export const TodoContext = createContext();
 
 export function TodoProvider({ children }) {
   const [todo, setTodo] = useState(
-    JSON.parse(localStorage.getItem("userTodoList")) || []
+    JSON.parse(localStorage.getItem("userTodoList")) || [
+      {
+        title: "Workout",
+        desc: "Today I am going to the gym for cardio and light weight training.",
+      },
+      {
+        title: "Learning React",
+        desc: "I will spend time revising the react hooks.and implement it.",
+      },
+      {
+        title: "Shopping",
+        desc: "Need to buy vegetables, fruits, and some daily essentials.",
+      },
+    ]
   );
   const [add, setAdd] = useState(false); // floating add button
   const [editId, setEditId] = useState(null); // access index by clicking edit button
   const [edit, setEdit] = useState(false); // toggle edit component
-  const [openMenu,setOpenMenu] = useState(false)
-   const [userEdit, setUserEdit] = useState({
+  const [openMenu, setOpenMenu] = useState(false);
+  const [userEdit, setUserEdit] = useState({
     title: "",
     desc: "",
   });
-  
-  const menuHandler=(e)=> {
-    e.stopPropagation(); 
-    setOpenMenu((prev) => !prev)
-  }
 
-  
+  const menuHandler = (e) => {
+    e.stopPropagation();
+    setOpenMenu((prev) => !prev);
+  };
+
   const editHandler = (id) => {
     setEdit((prev) => !prev);
-    setUserEdit(todo[id])
+    setUserEdit(todo[id]);
     setEditId(id);
   };
 
-  const deleteHandler = (id) => {
-    const updateTodo = todo.toSpliced(id, 1);
-    localStorage.setItem("userTodoList", JSON.stringify(updateTodo));
-    setTodo(updateTodo);
+  const deleteHandler = (id, todo) => {
+    const confirmation = confirm(`are you sure to delete ${todo.title}`);
+    if (confirmation) {
+      const updateTodo = todo.toSpliced(id, 1);
+      localStorage.setItem("userTodoList", JSON.stringify(updateTodo));
+      setTodo(updateTodo);
+      return;
+    }
   };
   const addHandler = () => {
     setAdd((prev) => !prev);
-    console.log("hi i am activate");
   };
 
   const favHandler = (index) => {
@@ -51,8 +66,6 @@ export function TodoProvider({ children }) {
       localStorage.setItem("userTodoList", JSON.stringify(todoUpdate));
     }
   };
-
-
 
   return (
     <TodoContext.Provider
@@ -75,7 +88,6 @@ export function TodoProvider({ children }) {
         setUserEdit,
       }}
     >
-
       {children}
     </TodoContext.Provider>
   );
